requirements.txt
fastapi
uvicorn
pydantic
email-validator
alembic
sqlalchemy
bcrypt>=4.0.0,<4.1.0
python-jose[cryptography]
python-multipart
passlib[bcrypt]>=1.7.4
PyJWT
pytest
pytest-mock
pytest-cov
python-decouple
paho-mqtt>=2.0.0